+++ idalocalgameai/__init__.py
@@ -1,4 +1,4 @@
 """MonsteyAI-IDA-plugin package."""
 
 PLUGIN_NAME = "MonsteyAI-IDA-plugin"
-PLUGIN_VERSION = "0.3.15"
+PLUGIN_VERSION = "0.3.16"